auth-btn-func
