+++ src/index.js
@@ -1,6 +1,6 @@
 import './sass/main.scss';
 import './js/api';
-import './js/authlib_btn';
+import './js/main';
 import './js/test';
 import './js/card_markup';
 import './js/render_trends';
