--- src/js/authlib_btn.js
@@ -1,23 +1,0 @@
-
-
-
-
-const libAuthBtnEl = document.querySelector('.auth_lib');
-
-
-// console.log(libAuthBtnEl);
-
-const isAuthenticated = false;
-
-// const isAuthenticated = true;
-
-if (isAuthenticated) {
-  libAuthBtnEl.classList.remove('auth');
-} else {
-  libAuthBtnEl.classList.add('auth');
-}
-
-
-
-
-
