+++ src/js/main.js
@@ -1,0 +1,5 @@
+import fetchPopularMovies from './render_trends';
+import onAuth from './onAuth';
+
+window.addEventListener('load', fetchPopularMovies);
+window.addEventListener('load', onAuth);
