+++ src/partials/header_library.html
@@ -4,7 +4,7 @@
       <div class="btn_group">
         <a class="btn btn--bigger" href="./index.html">Home</a>
         <button class="btn btn--current">My Library</button>
-        <button class="btn auth_lib" id="library-auth-btn"></button>
+        <button class="btn auth_lib auth" id="library-auth-btn"></button>
       </div>
       <div class="logo_position"> <a href="./index.html" class="logo_lib">Filmoteka</a></div>
     </div>
